requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
alembic==1.13.0
psycopg2-binary==2.9.9
youtube-transcript-api==0.6.1
google-api-python-client==2.108.0
openai==1.3.7
python-dotenv==1.0.0
schedule==1.2.0
python-telegram-bot==20.7
requests==2.31.0
beautifulsoup4==4.12.2
pandas==2.1.4
numpy==1.25.2
matplotlib==3.8.2
seaborn==0.13.0
jinja2==3.1.2
aiofiles==23.2.0
python-multipart==0.0.6
pydantic==2.5.0 